update code 18/12

+++ backend/routers/bill.router.js
@@ -228,7 +228,9 @@ router.get("/vnpay_return", async (req, res) => {
             const newBill = new Bill(dataBill);
 
             newBill.dataProduct.map(async (item) => {
-                const deleteCartById = await Cart.findByIdAndDelete(item.idCart);
+                if (item.idCart) {
+                    await Cart.findByIdAndDelete(item.idCart);
+                }
             })
 
             await newBill.save();
